moved code to bookgsrfragment.kt
